D100-D102: 200+ project improvements — errno hardening, static marking, code safety
Batch A: return -1 → errno in memory subsystem (217 fixes across 14 files)
Batch B: return -1 → errno in kernel/ (ftrace.c, 10 fixes)
Batch C: return -1 → errno in fs/process/drivers/net (17 files, ~900 fixes)
Batch D: signed 1<<N → 1U<<N (40+ fixes in 13 files + cascading sign-compare)
Batch E: switch default: cases (verified ~50, added 1 real fix in iosched.c)
Batch F: static marking in lib/*.c (34 functions, 16 files)
Batch G: division-by-zero guards (3 fixes in tcp_westwood/hybla)
Batch H: trailing whitespace cleanup (12 files, 303 lines)
Batch I: static marking in kernel/drivers/fs/net/process (67 functions, 18 files)

187 files changed, 1761 insertions(+), 1675 deletions(-)
Zero-warning build with -Wall -Wextra -Werror

+++ src/kernel/psi.c
@@ -32,7 +32,7 @@
  * ═══════════════════════════════════════════════════════════════════════ */
 
 #define PSI_FRAC_BITS  12          /* fixed-point resolution: 1/4096 */
-#define PSI_FRAC_ONE   (1 << PSI_FRAC_BITS)  /* 1.0 in fixed-point  */
+#define PSI_FRAC_ONE   (1U << PSI_FRAC_BITS)  /* 1.0 in fixed-point  */
 #define PSI_FRAC_HALF  (PSI_FRAC_ONE / 2)    /* 0.5 for rounding    */
 
 /* Format a fixed-point percentage value into a string like "12.34".
@@ -144,8 +144,8 @@ static void psi_update_window(struct psi_window *win,
     /* Clamp to [0, FRAC_ONE] */
     if (win->avg_fp < 0)
         win->avg_fp = 0;
-    if (win->avg_fp > PSI_FRAC_ONE)
-        win->avg_fp = PSI_FRAC_ONE;
+    if (win->avg_fp > (int)PSI_FRAC_ONE)
+        win->avg_fp = (int)PSI_FRAC_ONE;
 }
 
 /* ═══════════════════════════════════════════════════════════════════════
